+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- code cleanup
+
 ## [0.9.1] - 2025-10-31
 
 - fix missing types